add span test
